fix(ladybug): increase speed for query
